requirements-integration.txt
# Integration test dependencies with Testcontainers
testcontainers>=3.7.0
testcontainers-postgres>=0.0.1rc1
testcontainers-redis>=0.0.1rc1

# Python Testcontainers modules
docker>=6.1.0
requests>=2.31.0

# Additional test utilities for integration testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-timeout>=2.1.0
pytest-docker>=2.0.0

# Database clients for testing
psycopg2-binary>=2.9.0
redis>=4.5.0
pymongo>=4.4.0

# Vector store dependencies (for integration tests)
chromadb>=0.4.22
weaviate-client>=3.25.0
qdrant-client>=1.7.0
pymilvus>=2.3.0